DockerShenanigans/service/server/requirements.txt
grpcio==1.40.0
protobuf==3.20
dicttoxml==1.7.16
xmltodict==0.13.0
avro==1.11.1
msgpack==1.0.5
pyyaml==5.4.1
